backend/requirements.txt
pytest
PyPDF2
anthropic
python-dotenv
fastapi
uvicorn
python-multipart